+++ pkg/webserver/enqueue.go
@@ -1,4 +1,4 @@
-package main
+package webserver
 
 import (
 	"time"
