+++ test_run.sh
@@ -6,13 +6,9 @@ mkdir ./bin
 rm -f ./bin/broker
 go build -o ./bin/broker ./cmd/broker
 
-# Setup webserver
-rm -f ./bin/webserver
-go build -o ./bin/webserver ./cmd/webserver
-
 # Setup render
 rm -f ./bin/render
 go build -o ./bin/render ./cmd/render
 
 # Run
-PORT=3000 ./bin/webserver & ./bin/broker && fg
+PORT=3000 ./bin/broker && fg
